+++ site/dat/docs/Changelog.md
@@ -12,7 +12,7 @@
  - improve jmx2yaml conversion with encoded params
  - read errors info from distributed locustio test
  - avoid recursion in dummy screen dashboard
- 
+ - allow specifying link to download Grinder
 
 ## 1.9.1 <sup>19 apr 2017</sup>
  - fix errors in reading Grinder KPI file